requirements.txt
Django==1.6.5
Pillow==2.5.1
South==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-imagekit==3.2.2
django-suit==0.2.9
django-taggit==0.12
django-toolbelt==0.0.1
gunicorn==19.0.0
pilkit==1.1.9
psycopg2==2.5.3
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
